# Optimization: Execute Individual Sites in Parallel