Since the individual site algorithms are completely independent of each other, I made a slight change to run each site's computations in its own thread and then aggregate the results. On the example file, this leads to a 40% increase in performance (52s vs 72.5s average at N=1000); however, mileage will vary depending on how many sites there are.

### Performance discussion
In general, this will make files with multiple sites run significantly faster, but slightly slows down runs with just 1 site (since there's some overhead to starting a new thread). See the graph below (blue is the new multithreaded code, red is the current `master` branch):

![multiPerformanceGraph](https://user-images.githubusercontent.com/22411098/98743954-e2a13800-237e-11eb-8caa-8a705f79b7dc.PNG)

With 0 iterations, the serial code takes 1.41s and the multithreaded code 2.71s - a slight downgrade for low iteration counts or ~+1 second if there's only 1 site. Past ~50 iterations, though (well below what most researchers use), the example file runs significantly faster (theoretically approaching a 45% improvement as nruns goes to infinity *for this specific file*).

Since the 3rd site is a bottleneck in our example file (taking significantly longer), the gains aren't as significant as they could be. In a file with 5 sites each taking the same amount of time, this code would give nearly a 500% speed improvement.